+++ tests/.gitignore
@@ -1,0 +1,9 @@
+test
+
+# Generated by processor test
+trace*_edge_coverage.out
+trace*_bitmap.out
+
+# Generated by performance test
+execution_times_cache_mode.dat
+execution_times_non_cache_mode.dat
