+++ tests/branches/expected_edge_coverage.out
@@ -1,42 +1,42 @@
-0x900 [branches] -> 0x71c [branches]
-0x71c [branches] -> 0x728 [branches]
-0x728 [branches] -> 0x90c [branches]
-0x90c [branches] -> 0x734 [branches]
-0x734 [branches] -> 0x748 [branches]
-0x748 [branches] -> 0x748 [branches]
-0x748 [branches] -> 0x748 [branches]
-0x748 [branches] -> 0x754 [branches]
-0x754 [branches] -> 0x910 [branches]
-0x910 [branches] -> 0x764 [branches]
-0x764 [branches] -> 0x77c [branches]
-0x77c [branches] -> 0x794 [branches]
-0x794 [branches] -> 0x914 [branches]
-0x914 [branches] -> 0x7a8 [branches]
-0x7a8 [branches] -> 0x7c4 [branches]
-0x7c4 [branches] -> 0x7d8 [branches]
-0x7d8 [branches] -> 0x918 [branches]
-0x918 [branches] -> 0x7ec [branches]
-0x7ec [branches] -> 0x804 [branches]
-0x804 [branches] -> 0x81c [branches]
-0x81c [branches] -> 0x91c [branches]
-0x91c [branches] -> 0x830 [branches]
-0x830 [branches] -> 0x84c [branches]
-0x84c [branches] -> 0x860 [branches]
-0x860 [branches] -> 0x920 [branches]
-0x920 [branches] -> 0x874 [branches]
-0x874 [branches] -> 0x888 [branches]
-0x888 [branches] -> 0x880 [branches]
-0x880 [branches] -> 0x924 [branches]
-0x924 [branches] -> 0x898 [branches]
-0x898 [branches] -> 0x89c [branches]
-0x89c [branches] -> 0x8a0 [branches]
-0x8a0 [branches] -> 0x928 [branches]
-0x928 [branches] -> 0x8a8 [branches]
-0x8a8 [branches] -> 0x92c [branches]
-0x92c [branches] -> 0x8b8 [branches]
-0x8b8 [branches] -> 0x8c4 [branches]
-0x8c4 [branches] -> 0x930 [branches]
-0x930 [branches] -> 0x8d4 [branches]
-0x8d4 [branches] -> 0x8ec [branches]
-0x8ec [branches] -> 0x8e4 [branches]
-0x8e4 [branches] -> 0x934 [branches]
+0x900 [0] -> 0x71c [0]
+0x71c [0] -> 0x728 [0]
+0x728 [0] -> 0x90c [0]
+0x90c [0] -> 0x734 [0]
+0x734 [0] -> 0x748 [0]
+0x748 [0] -> 0x748 [0]
+0x748 [0] -> 0x748 [0]
+0x748 [0] -> 0x754 [0]
+0x754 [0] -> 0x910 [0]
+0x910 [0] -> 0x764 [0]
+0x764 [0] -> 0x77c [0]
+0x77c [0] -> 0x794 [0]
+0x794 [0] -> 0x914 [0]
+0x914 [0] -> 0x7a8 [0]
+0x7a8 [0] -> 0x7c4 [0]
+0x7c4 [0] -> 0x7d8 [0]
+0x7d8 [0] -> 0x918 [0]
+0x918 [0] -> 0x7ec [0]
+0x7ec [0] -> 0x804 [0]
+0x804 [0] -> 0x81c [0]
+0x81c [0] -> 0x91c [0]
+0x91c [0] -> 0x830 [0]
+0x830 [0] -> 0x84c [0]
+0x84c [0] -> 0x860 [0]
+0x860 [0] -> 0x920 [0]
+0x920 [0] -> 0x874 [0]
+0x874 [0] -> 0x888 [0]
+0x888 [0] -> 0x880 [0]
+0x880 [0] -> 0x924 [0]
+0x924 [0] -> 0x898 [0]
+0x898 [0] -> 0x89c [0]
+0x89c [0] -> 0x8a0 [0]
+0x8a0 [0] -> 0x928 [0]
+0x928 [0] -> 0x8a8 [0]
+0x8a8 [0] -> 0x92c [0]
+0x92c [0] -> 0x8b8 [0]
+0x8b8 [0] -> 0x8c4 [0]
+0x8c4 [0] -> 0x930 [0]
+0x930 [0] -> 0x8d4 [0]
+0x8d4 [0] -> 0x8ec [0]
+0x8ec [0] -> 0x8e4 [0]
+0x8e4 [0] -> 0x934 [0]
